Move interact logic into individual layer. Will only have InteractableLyric and lots of layer eventually.
