+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Content/Components/Lyrics/EditableLyric.cs
@@ -6,24 +6,13 @@ using System.Linq;
 using osu.Framework.Allocation;
 using osu.Framework.Extensions.IEnumerableExtensions;
 using osu.Framework.Graphics;
-using osu.Framework.Input.Events;
-using osu.Game.Rulesets.Karaoke.Edit.ChangeHandlers.Lyrics;
 using osu.Game.Rulesets.Karaoke.Objects;
-using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.CaretPosition;
-using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.CaretPosition.Algorithms;
-using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.States;
 
 namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Content.Components.Lyrics;
 
 [Cached(typeof(IEditableLyricState))]
 public partial class EditableLyric : InteractableLyric, IEditableLyricState
 {
-    [Resolved]
-    private ILyricsChangeHandler lyricsChangeHandler { get; set; } = null!;
-
-    [Resolved]
-    private ILyricCaretState lyricCaretState { get; set; } = null!;
-
     public EditableLyric(Lyric lyric)
         : base(lyric)
     {
@@ -35,6 +24,7 @@ public partial class EditableLyric : InteractableLyric, IEditableLyricState
     {
         return new Layer[]
         {
+            new EditLyricLayer(lyric),
             new TimeTagLayer(lyric),
             new CaretLayer(lyric),
             new BlueprintLayer(lyric),
@@ -44,116 +34,5 @@ public partial class EditableLyric : InteractableLyric, IEditableLyricState
     public void TriggerDisallowEditEffect()
     {
         InternalChildren.OfType<Layer>().ForEach(x => x.TriggerDisallowEditEffect(BindableMode.Value));
-    }
-
-    #region Hover
-
-    protected override bool OnMouseMove(MouseMoveEvent e)
-    {
-        if (!lyricCaretState.CaretEnabled)
-            return false;
-
-        // should not trigger the hover caret event if dragging other lyric.
-        if (e.HasAnyButtonPressed)
-            return false;
-
-        object? caretIndex = getCaretIndexByPosition(e);
-
-        if (caretIndex != null)
-        {
-            lyricCaretState.MoveHoverCaretToTargetPosition(Lyric, caretIndex);
-        }
-        else if (lyricCaretState.CaretPosition is not IIndexCaretPosition)
-        {
-            // still need to handle the case with non-index caret position.
-            lyricCaretState.MoveHoverCaretToTargetPosition(Lyric);
-        }
-
-        return base.OnMouseMove(e);
-    }
-
-    protected override void OnHoverLost(HoverLostEvent e)
-    {
-        base.OnHoverLost(e);
-
-        if (!lyricCaretState.CaretEnabled)
-            return;
-
-        // lost hover caret and time-tag caret
-        lyricCaretState.ClearHoverCaretPosition();
-    }
-
-    private object? getCaretIndexByPosition(UIEvent mouseEvent)
-    {
-        var algorithm = lyricCaretState.CaretPositionAlgorithm;
-        float xPosition = ToLocalSpace(mouseEvent.ScreenSpaceMousePosition).X;
-        return algorithm switch
-        {
-            CuttingCaretPositionAlgorithm => KaraokeSpriteText.GetCharIndicatorByPosition(xPosition),
-            TypingCaretPositionAlgorithm => KaraokeSpriteText.GetCharIndicatorByPosition(xPosition),
-            NavigateCaretPositionAlgorithm => null,
-            CreateRubyTagCaretPositionAlgorithm => KaraokeSpriteText.GetCharIndexByPosition(xPosition),
-            CreateRemoveTimeTagCaretPositionAlgorithm => KaraokeSpriteText.GetCharIndexByPosition(xPosition),
-            _ => null,
-        };
-    }
-
-    #endregion
-
-    #region Click
-
-    protected override bool OnClick(ClickEvent e)
-    {
-        return lyricCaretState.ConfirmHoverCaretPosition();
-    }
-
-    #endregion
-
-    #region Drag
-
-    protected override bool OnDragStart(DragStartEvent e)
-    {
-        // should not handle the drag event if the caret algorithm is able to handle it.
-        if (!lyricCaretState.CaretDraggable)
-            return false;
-
-        // confirm the hover caret position before drag start.
-        return lyricCaretState.StartDragging();
-    }
-
-    protected override void OnDrag(DragEvent e)
-    {
-        object? caretIndex = getCaretIndexByPosition(e);
-
-        if (caretIndex != null)
-        {
-            lyricCaretState.MoveDraggingCaretIndex(caretIndex);
-        }
-
-        base.OnDrag(e);
-    }
-
-    protected override void OnDragEnd(DragEndEvent e)
-    {
-        lyricCaretState.EndDragging();
-
-        base.OnDragEnd(e);
-    }
-
-    #endregion
-
-    protected override bool OnDoubleClick(DoubleClickEvent e)
-    {
-        var position = lyricCaretState.CaretPosition;
-
-        switch (position)
-        {
-            case CuttingCaretPosition cuttingCaretPosition:
-                lyricsChangeHandler.Split(cuttingCaretPosition.CharGap);
-                return true;
-
-            default:
-                return false;
-        }
     }
 }
