+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Content/Components/Lyrics/ViewOnlyLyric.cs
@@ -2,18 +2,12 @@
 // See the LICENCE file in the repository root for full licence text.
 
 using System.Collections.Generic;
-using osu.Framework.Allocation;
-using osu.Framework.Input.Events;
 using osu.Game.Rulesets.Karaoke.Objects;
-using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.States;
 
 namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Content.Components.Lyrics;
 
 public partial class ViewOnlyLyric : InteractableLyric
 {
-    [Resolved]
-    private ILyricCaretState lyricCaretState { get; set; } = null!;
-
     public ViewOnlyLyric(Lyric lyric)
         : base(lyric)
     {
@@ -23,33 +17,8 @@ public partial class ViewOnlyLyric : InteractableLyric
     {
         return new Layer[]
         {
+            new InteractLyricLayer(lyric),
             new TimeTagLayer(lyric),
         };
-    }
-
-    protected override bool OnMouseMove(MouseMoveEvent e)
-    {
-        if (!lyricCaretState.CaretEnabled)
-            return false;
-
-        if (IsDragged)
-            return false;
-
-        lyricCaretState.MoveHoverCaretToTargetPosition(Lyric);
-
-        return base.OnMouseMove(e);
-    }
-
-    protected override void OnHoverLost(HoverLostEvent e)
-    {
-        base.OnHoverLost(e);
-
-        // lost hover caret and time-tag caret
-        lyricCaretState.ClearHoverCaretPosition();
-    }
-
-    protected override bool OnClick(ClickEvent e)
-    {
-        return lyricCaretState.MoveCaretToTargetPosition(Lyric);
     }
 }
